requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# LangChain and LangGraph
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.26
langchain-core==0.1.3

# OpenAI
openai==1.6.1

# Utilities
colorama==0.4.6
requests==2.31.0
typing-extensions==4.9.0

# For graph visualization (optional)
pillow==10.1.0
graphviz==0.20.1